meas 0.1.7 is out. Intervall duration can be set as floating point value, if required. Intervall duation has a resulution of 1 ms. Smallest possible value 0.010 s

+++ sw/meas/src/types.h
@@ -151,7 +151,7 @@ typedef struct
 
   pid_t      pid;
 
-  uint16_t   interval_duration;
+  uint32_t   interval_duration;
 
   uint32_t   no_intervals;
